backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.0
bcrypt>=3.2.2,<4.0.0

# Cache
redis==5.0.1

# Scheduler
apscheduler==3.10.4

# HTTP Client
httpx==0.26.0

# Tools
python-dotenv==1.0.0
